Refactor Firebase helpers
